F4-Sprint2-TranscriptionSummarization - Add base64 utils
